third_party/WebKit/LayoutTests/http/tests/security/mixedContent/laxChecking/insecure-sync-post-xhr-blocked-expected.txt

ALERT: Executing script in the child window.
ALERT: XHR POST opened.
CONSOLE WARNING: line 7: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-sync-xhr-post.html' was loaded over HTTPS, but displayed insecure content from 'http://localhost:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*&headers=Content-Type': this content should also be loaded over HTTPS.

ALERT: XHR POST sent.
This test opens a HTTPS window that loads insecure data via XHR. We should trigger a mixed content callback and block the load.
